requirements/production.txt
-r base.txt

# ============================================
# WSGI Server
# ============================================
gunicorn==21.2.0

# ============================================
# Celery & Redis
# ============================================
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# ============================================
# Storage (for production)
# ============================================
boto3==1.34.10
django-storages==1.14.2

# ============================================
# Reports
# ============================================
reportlab==4.0.7
openpyxl==3.1.2
WeasyPrint==60.1

# ============================================
# Monitoring
# ============================================
sentry-sdk==1.39.1